Various patches are applied
Patch 9/9
	Fix the bug of ignoring updated content info.

	Even if the box tries to update its content info,
	in case of script type box, the provider ignores it.

	This patch will fix it.

Patch 8/9
	Use the break after handling the service request (badge/noti)

Patch 7/9
	Using EINA_LIST_FREE instead of EINA_LIST_FOREACH_SAFE.
	To release the svc_ctx->packet_list.

Patch 6/9
	Keep the desc block even if the buffer of lb is not ready.
	Keep the desc blocks even if the buffer of livebox is not created.
	The provider will send the created event to the slave.
	Then the slave will start to generate the desc file. (script data)
	But the provider cannot apply desc to its buffer.
	Because there is no created object.
	To resolve this issue,
	This patch will create a cache.
	It will be used for keeping the blocks which are parsed before creating a buffer of livebox (or PD).

Patch 5/9
	Update License

Patch 4/9
	Enable Badge/Notification service.

Patch 3/9
	Don't set the "default" to content as default value.
	Client can use "" for its content value.
	So the provider has not to touch the "content".

Patch 2/9
Update period compensation timer.

Patch 1/9
	Fix the storage size calculation bug.

+++ include/util.h
@@ -1,7 +1,7 @@
 /*
  * Copyright 2013  Samsung Electronics Co., Ltd
  *
- * Licensed under the Flora License, Version 1.0 (the "License");
+ * Licensed under the Flora License, Version 1.1 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
@@ -22,7 +22,7 @@ extern int util_unlink(const char *filename);
 extern int util_unlink_files(const char *folder);
 extern char *util_slavename(void);
 extern const char *util_basename(const char *name);
-extern unsigned long util_free_space(const char *path);
+extern unsigned long long util_free_space(const char *path);
 extern char *util_replace_string(const char *src, const char *pattern, const char *replace);
 extern const char *util_uri_to_path(const char *uri);
 extern void *util_timer_add(double interval, Eina_Bool (*cb)(void *data), void *data);
